backend/pyproject.toml
[tool.poetry]
name = "ai-ceo-clone-backend"
version = "0.1.0"
description = "AI CEO Clone Interface Backend"
authors = ["Devin AI <devin@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
openai = "^1.3.0"
elevenlabs = "^0.2.26"
aiofiles = "^23.2.1"
python-magic = "^0.4.27"
pypdf2 = "^3.0.1"
python-docx = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3